added Model, need to work on authentication to backend
