--- client/src/App.css
@@ -1,11 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.input-group{
-  background-color:black;
-  width:100px;
-}
-.sign-in-container{
-  width:100px;
-}
