+++ client/src/SignIn.css
@@ -1,0 +1,15 @@
+.App {
+  text-align: center;
+}
+
+.input-group{
+  background-color:black;
+  width:100px;
+}
+.sign-in-container{
+  width:100px;
+}
+
+h3{
+  margin-bottom:10px;
+}
