+++ client/src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import {createStore } from 'redux';
@@ -8,7 +7,7 @@ import {mainReducer} from './reducers/main-reducer.js'
 import {Provider} from 'react-redux';
 import Bootstrap from 'bootstrap/dist/css/bootstrap.css';
 import {BrowserRouter as Router} from 'react-router-dom';
-
+import './index.css';
 let store = createStore(mainReducer);
 
 ReactDOM.render(<Provider store={store}><App /></Provider>, document.getElementById('root'));
